odc_index/requirements.txt
aiobotocore[awscli,boto3]
datacube
odc-index
odc-aws
odc-aio
odc-apps-cloud